Incorporated the gridcell ARIA role into the code.
The detection of an element with a `role` attribute with a value of `gridcell` now enables directional navigation. The detection of an element with a `tabindex` attribute no longer enables directional navigation. This is because it is very vague to assume that `tabdindex` on its own indicates the presence of a composite widget.

As a side note, `KeyboardEvent.keyCode` is no longer used. It has been deprecated. `KeyboardEvent.key` is used instead.