refactor: Use focusin, focusout and pointerdown events in remove event remove
Before, the blur, click and focus events were used by the remove event
component. Now, the focusin event replaces the focus event; the focusout
event replaces the blur event; and the pointerdown event replaces the
click event. The purpose of this change is to align the code with the
latest API's.