refactor: Visually represent a file object in HTML
Before, an event called 'onDidAddFile' was fired. An event handler
listening for this event was in charge of generating the HTML to visually
represent a file object. Now, the notion of an event being fired is no
longer in use. A plain function object that generates this representation
is used instead.