refactor: Map file objects in HTML to file select controls
Before, file select controls mapped to another map of file objects in HTML
to a list of file objects. This implementation was not compatible with the
logic that handles the removal of a file object in HTML. Now, there are
two separate maps: file objects in HTML to file select controls and file
select controls to a list of file objects.