refactor: Implement prototypal inheritance
Before, this file consisted only of private functions and variables. This
made it impossible to interact with them from an outer scope. Now, some of
these functions and variables are available through the prototype property
of FileSelectControl which is now meant to be used as a constructor.
Additionally, some identifiers have been changed to make the code more
readable. JSDoc comments have also been updated accordingly.