fix: Stop searching for a next or previous sibling if there is no parent.
Before, it was possible for the root node to retrieve a next or previous
sibling node. This is in fact useless as the root node (by definition) has
no parent or siblings. Trying to retrieve a next or previous sibling
resulted in a JavaScript error. Now, the code simply returns prematurely
if the node in question has no parent.