feat: Create the ability to build a menu
The following piece of code is in charge of building a widget that offers
a list of choices to the user, such as a set of actions or functions. This
is known as a menu. A menu consists of one or more menu items. A menu may
be complex in that it may include sub-level menus which in turn own one or
more menu items.