feat: Check whether a node has children
Before, the only way to verify if a node had children was through the
'length' property of the same node's 'childNodes' property. For example,
'node.childNodes.length'. This was a bit cumbersome. Now, a new method has
been added to the same node's prototype to verify if this same node has
children or not. For example, 'node.hasChildNodes()'.