chore: Generate most basic representation of a menubar in HTML markup
Before, it was very tedious and repetitive to structure a menubar in HTML
markup. Now, there is a piece of code that generates the most basic
representation of a menubar in HTML markup. This representation resembles
a tree data structure. The HTML markup may be modified afterwards to
customize the menubar.