refactor: Exclude context node from result returned by `find` and `findAll`
Before, the context node could be returned by `Node.prototype.find` and
`Node.prototype.findAll`. This happened if the context node passed the
testing function provided to both methods. Now, the context node is
skipped and ignored. For that reason, it will be excluded from the result
returned by both methods.