refactor: Hide the underlying data an instance of Queue operates on
Before, it was possible for external code to access the underlying data of a
queue. This data was in the form of an array. It was accessible through the
`data` property of a queue. Now, external code cannot access the underlying
data. This is possible through the use of `WeakMap`. Additionally, comments
have been updated.