refactor: Apply better formatting and best practices to this piece of code
Before, this piece of code did not comply with what are considered best practices in JavaScript. For example, loose equality was used instead of strict equality. Now, this piece of code has been refactored in such a way that it is aligned with how other pieces of code have been written in this repository. Also, JSDOC comments have been added.